Hack beta1 to prevent inexplicable initialization error
* Make update_decay only do beta1 assign if beta1_during_...
is not the same as beta1
* Make beta1 not a variable in build_graph but an actual
value

+++ vat_ladder/src/utils.py
@@ -33,7 +33,7 @@ def get_cli_params():
     # EVALUATE
     # -------------------------
 
-    add('--test_frequency_in_epochs', default=5, type=int)
+    add('--test_frequency_in_epochs', default=5, type=float)
     # validation
     add('--validation', default=0, nargs='?', const=1000, type=int)
